+++ src/main/java/com/aparapi/internal/kernel/KernelPreferences.java
@@ -27,7 +27,7 @@ import java.util.concurrent.atomic.AtomicReference;
 public class KernelPreferences {
    private final Class<? extends Kernel> kernelClass;
    private final KernelManager manager;
-   private final AtomicReference<LinkedList<Device>> preferredDevices = new AtomicReference<>(null);
+   private final AtomicReference<LinkedHashSet<Device>> preferredDevices = new AtomicReference<>(null);
    private final LinkedHashSet<Device> failedDevices = new LinkedHashSet<>();
 
    public KernelPreferences(KernelManager manager, Class<? extends Kernel> kernelClass) {
@@ -60,6 +60,24 @@ public class KernelPreferences {
       }
       return Collections.unmodifiableList(localPreferredDevices);
    }
+   
+   /**
+    * Validates if the specified devices is among the preferred devices for executing the kernel associated with the current
+    * kernel preferences.
+    * @param device the device to be tested
+    * @return <ul><li>true, if specified device is among the preferred devices</li>
+    *             <li>false, otherwise</li></ul>
+    */
+   public boolean isDeviceAmongPreferredDevices(Device device) {
+	   maybeSetUpDefaultPreferredDevices();
+	   
+	   boolean result = false;
+	   synchronized (this) {
+		   result = preferredDevices.get().contains(device);
+	   }
+	   
+	   return result;
+   }
 
    synchronized void setPreferredDevices(LinkedHashSet<Device> _preferredDevices) {
       if (preferredDevices.get() != null) {
@@ -67,7 +85,7 @@ public class KernelPreferences {
          preferredDevices.get().addAll(_preferredDevices);
       }
       else {
-         preferredDevices.set(new LinkedList<>(_preferredDevices));
+         preferredDevices.set(new LinkedHashSet<>(_preferredDevices));
       }
       failedDevices.clear();
    }
@@ -78,14 +96,23 @@ public class KernelPreferences {
    }
 
    synchronized void markPreferredDeviceFailed() {
-      if (preferredDevices.get().size() > 0) {
-         failedDevices.add(preferredDevices.get().remove(0));
+	  LinkedHashSet<Device> devices = preferredDevices.get();
+      if (devices.size() > 0) {
+    	 Device device = devices.iterator().next();
+    	 preferredDevices.get().remove(device);
+         failedDevices.add(device);
       }
    }
+   
+   synchronized void markDeviceFailed(Device device) {
+	   preferredDevices.get().remove(device);
+       failedDevices.add(device);
+   }
+
 
    private void maybeSetUpDefaultPreferredDevices() {
 	   if (preferredDevices.get() == null) {
-		   preferredDevices.compareAndSet(null, new LinkedList<>(manager.getDefaultPreferences().getPreferredDevices(null)));
+		   preferredDevices.compareAndSet(null, new LinkedHashSet<>(manager.getDefaultPreferences().getPreferredDevices(null)));
 	   }
    }